+++ Sources/Network/APIConstants.swift
@@ -16,7 +16,7 @@ struct APIConstants {
     static let createPostURL = baseURL + "/post"
     static let plusVoteNumbersURL = baseURL + "/post"
     static let findAllBestPostURL = baseURL + "/post/list"
-    static let detailPostURL = baseURL + "/post"
+    static let detailPostURL = baseURL + "/post/"
     static let deletePostURL = baseURL + "/post/"
     static let postImageUploadURL = baseURL + "/image"
