+++ Sources/Present/Common/Model/VoteModel.swift
